+++ app/javascript/packs/application.js
@@ -4,10 +4,10 @@
 // that code so it'll be compiled.
 
 require("@rails/ujs").start()
-require("turbolinks").start()
 require("@rails/activestorage").start()
 require("channels")
-require('./preview')
+require('jquery')
+require("../preview")
 import "bootstrap"
 import "../stylesheets/application"
