+++ package.json
@@ -8,6 +8,7 @@
     "@rails/ujs": "^6.0.0-alpha",
     "@rails/webpacker": "4.3.0",
     "bootstrap": "^5.2.3",
+    "jquery": "^3.6.4",
     "turbolinks": "^5.2.0"
   },
   "version": "0.1.0",
